+++ Assets/Scripts/Business/Reactive/EarnedMoneyEvent.cs
@@ -3,6 +3,6 @@
     /// Event that being called after one earn tick
     public struct EarnedMoneyEvent
     {
-        public int moneyToAdd;
+        public ulong moneyToAdd;
     }
 }
