+++ Assets/Scripts/Utils/Literals.cs
@@ -5,10 +5,10 @@
     {
         public const string SaveFileName = "Saves";
         
-        public static string GetCostLabel(int value) => $"Cost: {value}$";
-        public static string GetPriceLabel(int value) => $"{value}$";
+        public static string GetCostLabel(ulong value) => $"Cost: {value}$";
+        public static string GetPriceLabel(ulong value) => $"{value}$";
         public static string GetPurchasedLabel() => "Purchased";
-        public static string GetBalanceLabel(int value) => $"Balance: {value}$";
+        public static string GetBalanceLabel(ulong value) => $"Balance: {value}$";
         public static string GetEarnMultiplierLabel(int value) => $"Earn: {value}%";
         public static string GetSpeedLabel(int value) => $"X{value}";
     }
